style: rename user.model
